pyproject.toml
[tool.poetry]
name = "trampolines"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.0"
matplotlib = "^3.5.1"
Pillow = "^9.0.0"
keras = "^2.7.0"
requests = "^2.27.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
